feat: implement CatalogScreen and enhance HomeScreen with promotional banner and search functionality
