+++ frontend/src/app/catalog/page.tsx
@@ -1,0 +1,5 @@
+import { CatalogScreen } from "@/features/catalog/screens/CatalogScreen";
+
+export default function Page() {
+  return <CatalogScreen />;
+}
