+++ frontend/src/shared/ui/Header.module.css
@@ -3,6 +3,13 @@
   top: 0;
   z-index: 10;
   background: var(--header-bg);
+  overflow: visible;
+}
+
+.header[data-search-open="true"] {
+  background:
+    radial-gradient(circle at 18% 18%, rgb(83 93 223 / 0.24) 0, transparent 32%),
+    linear-gradient(180deg, #131417 0%, #0b0f22 40%, #08083d 100%);
 }
 
 .inner {
@@ -65,9 +72,10 @@
 
 .search input {
   height: var(--header-search-height);
-  border-color: var(--color-white);
+  border-color: rgb(124 144 224 / 0.85);
   border-radius: var(--header-search-radius);
-  background: var(--header-bg);
+  background: #050508;
+  box-shadow: 0 0 0 1px rgb(83 93 223 / 0.18);
 }
 
 .search input::placeholder {
@@ -80,8 +88,227 @@
 
 .search input:focus {
   border-color: var(--color-primary-blue);
+  box-shadow:
+    0 0 0 1px rgb(83 93 223 / 0.32),
+    0 0 0 3px rgb(83 93 223 / 0.12);
 }
 
 .search svg {
   color: var(--color-white);
+}
+
+.searchSheet {
+  position: fixed;
+  inset: 0;
+  z-index: 20;
+  display: none;
+  overflow: hidden;
+  overscroll-behavior: none;
+  touch-action: none;
+  color: var(--color-text);
+  background:
+    radial-gradient(circle at 20% 14%, rgb(83 93 223 / 0.28) 0, transparent 28%),
+    linear-gradient(180deg, #171622 0%, #0b1026 38%, #08083d 100%);
+}
+
+.searchSheetInner {
+  position: relative;
+  z-index: 1;
+  display: grid;
+  align-content: start;
+  gap: var(--space-4);
+  height: 100%;
+  max-height: 100dvh;
+  overflow: hidden;
+  padding: 30px 20px 24px;
+}
+
+.searchSheetHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--space-8);
+}
+
+.searchTitle {
+  margin: 0;
+  font-family: var(--font-primary);
+  font-size: 20px;
+  font-weight: 600;
+  line-height: 1.1;
+  color: var(--color-white);
+}
+
+.closeButton {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: var(--header-icon-action-size);
+  height: var(--header-icon-action-size);
+  padding: 0;
+  border: 0;
+  border-radius: var(--radius-sm);
+  background: transparent;
+  color: var(--color-white);
+  cursor: pointer;
+}
+
+.closeButton:hover {
+  opacity: 0.8;
+}
+
+.closeButton:focus-visible {
+  outline: 2px solid var(--color-primary-blue);
+  outline-offset: 2px;
+}
+
+.previewCard {
+  display: grid;
+  gap: var(--space-12);
+  margin-top: 4px;
+  padding: var(--space-12);
+  border-radius: var(--radius-sm);
+  background: #050508;
+}
+
+.previewHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--space-8);
+  color: var(--color-white);
+}
+
+.previewTitle {
+  font-size: 16px;
+  font-weight: 500;
+  line-height: var(--typography-body1-line-height);
+}
+
+.previewList {
+  list-style: none;
+  display: grid;
+  gap: var(--space-8);
+  margin: 0;
+  padding: 0;
+}
+
+.emptyState {
+  margin: 0;
+  font-size: 16px;
+  line-height: 1.4;
+  color: #c0c1c6;
+  font-family: var(--font-primary);
+  font-style: italic;
+}
+
+.loadingState {
+  display: flex;
+  align-items: center;
+  gap: var(--space-12);
+  padding-block: var(--space-4);
+}
+
+.spinner {
+  width: 20px;
+  height: 20px;
+  flex-shrink: 0;
+  border-radius: 9999px;
+  border: 2px solid rgb(255 255 255 / 0.2);
+  border-top-color: var(--color-primary-pink);
+  animation: searchSpin 0.9s linear infinite;
+}
+
+.loadingText {
+  font-size: 16px;
+  line-height: 1.4;
+  color: #c0c1c6;
+  font-family: var(--font-primary);
+  font-style: italic;
+}
+
+@keyframes searchSpin {
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+.previewItem {
+  display: flex;
+  align-items: baseline;
+  justify-content: space-between;
+  gap: var(--space-12);
+  width: 100%;
+  font-size: var(--typography-body2-font-size);
+  line-height: 1.4;
+  color: #c0c1c6;
+  text-decoration: none;
+  transition: opacity var(--ui-transition-fast);
+}
+
+.previewItem:hover {
+  opacity: 0.8;
+}
+
+.previewName {
+  min-width: 0;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  font-size: 16px;
+  color: #c0c1c6;
+  font-family: var(--font-primary);
+  font-style: italic;
+}
+
+.previewPrice {
+  flex-shrink: 0;
+  font-size: 16px;
+  color: #c0c1c6;
+  font-family: var(--font-primary);
+  font-style: italic;
+}
+
+.cat {
+  position: absolute;
+  right: -10px;
+  top: 50%;
+  width: 124px;
+  height: auto;
+  transform: translateY(-50%);
+  pointer-events: none;
+  filter: drop-shadow(0 12px 24px rgb(0 0 0 / 0.28));
+}
+
+@media (min-width: 768px) {
+  .searchSheet {
+    display: none;
+  }
+}
+
+@media (max-width: 767px) {
+  .header[data-search-open="true"] .topRow {
+    display: none;
+  }
+
+  .header[data-search-open="true"] .primarySearch {
+    display: none;
+  }
+
+  .header[data-search-open="true"] .searchSheet {
+    display: grid;
+    touch-action: auto;
+  }
+
+  .header[data-search-open="true"] .searchSheetInner {
+    touch-action: manipulation;
+  }
+
+  .header[data-search-open="true"] .search {
+    width: 100%;
+  }
+
+  .header[data-search-open="true"] .sheetSearchField {
+    margin-top: 12px;
+  }
 }
